refactor(agent): hoist provisioning types out of api.ts into types.ts
SignedTurnkeyRequest, ProvisioningActivity, ProvisioningInitResponse,
and ProvisioningStepInput were declared inline next to the fetch
helpers; move them to lib/types.ts to match the convention used by the
rest of the file. useAgent.ts imports SignedTurnkeyRequest from there
directly now instead of re-exporting via api.ts.
