+++ lib/api.ts
@@ -76,8 +76,11 @@ import {
   MppCredentialsResponse,
   Points,
   PromotionsBannerResponse,
+  ProvisioningActivity,
+  ProvisioningInitResponse,
   ProvisioningSessionRequest,
   ProvisioningSessionResponse,
+  ProvisioningStepInput,
   RainConsumerType,
   RainContractResponseDto,
   RainKycSubmitResponse,
@@ -2558,22 +2561,6 @@ const agentJsonHeaders = () => {
   };
 };
 
-/**
- * Envelope returned by the Turnkey SDK's `stampX` methods. `body` is the
- * exact stringified bytes the SDK signed — we MUST forward it verbatim;
- * re-serializing on the server changes key order and breaks the stamp.
- */
-export type SignedTurnkeyRequest = {
-  url: string;
-  body: string;
-  stamp: { stampHeaderName: string; stampHeaderValue: string };
-};
-
-export type ProvisioningActivity = {
-  url: string;
-  body: Record<string, unknown>;
-};
-
 const postAgentJson = async <T>(path: string, body?: unknown): Promise<T> => {
   const response = await fetch(agentEndpoint(path), {
     method: 'POST',
@@ -2585,26 +2572,20 @@ const postAgentJson = async <T>(path: string, body?: unknown): Promise<T> => {
   return response.json();
 };
 
-export const provisionAgentInit = (): Promise<{
-  provisioningId: string;
-  activity: ProvisioningActivity;
-}> => postAgentJson('/provision/init');
+export const provisionAgentInit = (): Promise<ProvisioningInitResponse> =>
+  postAgentJson('/provision/init');
 
-export const provisionAgentWalletAccount = (input: {
-  provisioningId: string;
-  signed: SignedTurnkeyRequest;
-}): Promise<{ activity: ProvisioningActivity }> =>
-  postAgentJson('/provision/wallet-account', input);
+export const provisionAgentWalletAccount = (
+  input: ProvisioningStepInput,
+): Promise<{ activity: ProvisioningActivity }> => postAgentJson('/provision/wallet-account', input);
 
-export const provisionAgentUser = (input: {
-  provisioningId: string;
-  signed: SignedTurnkeyRequest;
-}): Promise<{ activity: ProvisioningActivity }> => postAgentJson('/provision/user', input);
+export const provisionAgentUser = (
+  input: ProvisioningStepInput,
+): Promise<{ activity: ProvisioningActivity }> => postAgentJson('/provision/user', input);
 
-export const provisionAgentPolicy = (input: {
-  provisioningId: string;
-  signed: SignedTurnkeyRequest;
-}): Promise<{ agentEoaAddress: string }> => postAgentJson('/provision/policy', input);
+export const provisionAgentPolicy = (
+  input: ProvisioningStepInput,
+): Promise<{ agentEoaAddress: string }> => postAgentJson('/provision/policy', input);
 
 export const fetchAgent = async (): Promise<AgentSummary> => {
   const response = await fetch(agentEndpoint('/me'), {
